% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sas_session}
\alias{get_sas_session}
\title{Get the last/default SAS session}
\usage{
get_sas_session()
}
\description{
Get the last/default SAS session
}
\details{
this function is designed to facilitate the R users programming practice
of function oriented programming instead of object oriented programmings.
}
