% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_saspy}
\alias{install_saspy}
\title{install \code{saspy} module}
\usage{
install_saspy(method = "auto", conda = "auto")
}
\arguments{
\item{method}{method to install \code{saspy}}

\item{conda}{whether to use conda}
}
\description{
install \code{saspy} module
}
