% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sas_session}
\alias{sas_session}
\title{Create SAS Session Based on Configuration File}
\usage{
sas_session(cfgfile = get_sas_cfg(), ...)
}
\arguments{
\item{cfgfile}{(\code{string})\cr SAS session configuration.}

\item{...}{additional arguments passed to \code{saspy.SASsession()}.
Can override the configuration file.}
}
\value{
SAS session.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Create a SAS session.
}
