% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sas_session_ssh}
\alias{sas_session_ssh}
\title{Create SAS ssh Session Based on Configuration File}
\usage{
sas_session_ssh(sascfg = get_sas_cfg())
}
\arguments{
\item{sascfg}{(\code{character})\cr SAS session configuration.}
}
\value{
SAS session.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Create a SAS ssh session.
}
