% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sascfg.R
\name{sascfg}
\alias{sascfg}
\title{create SAS session configuration file}
\usage{
sascfg(
  name = "default",
  host,
  sas,
  ssh = system("which ssh", intern = TRUE),
  encoding = "latin1",
  options = list("-fullstimer"),
  ...,
  sascfg = "sascfg_personal.py"
)
}
\arguments{
\item{name}{(\code{character})\cr name of the configuration.}

\item{host}{(\code{character})\cr host name of remote server.}

\item{sas}{(\code{character})\cr SAS executable path on remote server.}

\item{ssh}{(\code{character})\cr executable path of ssh.}

\item{encoding}{(\code{character})\cr encoding of the SAS session.}

\item{...}{additional arguments.}

\item{sascfg}{(\code{character})\cr target file of configuration.}
}
\description{
create SAS session configuration file
}
\details{
\code{host} and \code{sas} are required to connect to remote SAS server. Other arguments can follow default.
If transferring datasets is needed, then tunnelling is required.
Use \verb{tunnel = }, \verb{rtunnel = } to specify tunnels and reverse tunnels.
The values should be length 1 integer.
}
