% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{saspy}
\alias{saspy}
\title{saspy package}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 0.
}
\usage{
saspy
}
\description{
saspy package
}
\keyword{internal}
