% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sasr-package.R
\docType{package}
\name{sasr-package}
\alias{sasr-package}
\alias{_PACKAGE}
\title{\code{sasr} Package}
\description{
\code{sasr} provides interface to SAS through \code{saspy} and \code{reticulate} in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/insightsengineering/sasr/}
  \item Report bugs at \url{https://github.com/insightsengineering/sasr/issues}
}

}
\author{
\strong{Maintainer}: Liming Li \email{liming.li@roche.com}

Authors:
\itemize{
  \item Daniel Sabanes Bove \email{daniel.sabanes_bove@roche.com}
  \item Isaac Gravestock \email{isaac.gravestock@roche.com}
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
