% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sasr.R
\docType{package}
\name{sasr}
\alias{sasr}
\title{Connect to SAS® using the IOM Bridge}
\description{
The \strong{sasr} package contains functions to connect to 
a SAS® server over the IOM Bridge.  The IOM Bridge allows you to read and
write SAS® datasets, run SAS® code, and interact with the SAS® meta-data 
server.    

The \strong{sasr} package is part of the \code{\link[sassy]{sassy}} family 
of packages, which help bring SAS® concepts to R.
}
