% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_data}
\alias{validate_data}
\title{Validate and Process \code{data.frame} for SAS}
\usage{
validate_data(data)
}
\arguments{
\item{data}{(\code{data.frame})\cr data.frame to be checked.}
}
\value{
\code{data.frame}\cr
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Validate if data contains validate variable names in SAS, and
remove possible row names.
}
\details{
In SAS, the variable names should be consist of letters, numbers and underscore.
Other characters are not allowed.
In addition, in SAS, row names(index) are not allowed.
}
\keyword{internal}
