% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_sascfg}
\alias{validate_sascfg}
\title{Validate SAS Configuration File Exist}
\usage{
validate_sascfg(sascfg)
}
\arguments{
\item{sascfg}{(\code{character})\cr file path of configuration.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Validate if SAS configuration file exist.
}
\details{
Currently, only the file existence check is conducted and the rest
is checked at python side.
}
\keyword{internal}
