% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_ssh_with_tunnel}
\alias{validate_ssh_with_tunnel}
\title{validate the SAS ssh session is using tunnels}
\usage{
validate_ssh_with_tunnel(
  session,
  msg = "SAS session through ssh must use tunnels to transfer datasets!"
)
}
\arguments{
\item{session}{SAS session}

\item{msg}{(\code{character})\cr message to display}
}
\description{
validate the SAS ssh session is using tunnels
}
