% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sas.R
\name{df2sd}
\alias{df2sd}
\title{transfer data.frame to sas data}
\usage{
df2sd(df, table = "_df", libref = "", ..., sas_session = get_sas_session())
}
\arguments{
\item{df}{(\code{data.frame})\cr data frame to be transferred.}

\item{table}{(\code{character})\cr table name in SAS.}

\item{libref}{(\code{character})\cr library name in SAS.}

\item{sas_session}{(\code{saspy.sasbase.SASsession}) SAS session}
}
\description{
transfer data.frame to sas data
}
