% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sas_cfg}
\alias{get_sas_cfg}
\title{Obtain the SAS configuration file}
\usage{
get_sas_cfg()
}
\description{
Obtain the SAS configuration file
}
\details{
Obtain the default sas configuration file. By default, it will search
the \code{sascfg_personal.py} file under current directory. If it does not exist, it will
search this file under home directory. If this file does not exist, NULL will be returned.
}
