% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sas_session}
\alias{get_sas_session}
\title{Get the Last or Default SAS Session}
\usage{
get_sas_session()
}
\value{
SAS session.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Obtain the last session or default session.
}
\details{
this function is designed to facilitate the R users programming practice
of function oriented programming instead of object oriented programmings.
}
