% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_saspy}
\alias{install_saspy}
\title{Install \code{saspy} Module}
\usage{
install_saspy(method = "auto", conda = "auto")
}
\arguments{
\item{method}{(\code{character})\cr method to install \code{saspy}.}

\item{conda}{(\code{character})\cr path to \code{conda} executable.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Install \code{saspy} module in \code{reticulate}.
}
