% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sas.R
\name{run_sas}
\alias{run_sas}
\title{Run SAS Code with SAS Session}
\usage{
run_sas(sas_code, results = c("TEXT", "HTML"), sas_session = get_sas_session())
}
\arguments{
\item{sas_code}{(\code{character})\cr sas code to be executed.}

\item{results}{(\code{character})\cr sas code execution results type.}

\item{sas_session}{(\code{saspy.sasbase.SASsession}) SAS session.}
}
\value{
a list with name \code{LST} and \code{LOG}.
}
\description{
Run SAS Code with SAS Session
}
\details{
\code{run_sas} will run sas code through SAS session.
The results is a named list of \code{LST} and \code{LOG}.
The result part will be stored in \code{LST}, and log will be stored in \code{LOG}.
If \code{results} argument is "TEXT", then results are in text format;
if \code{results} argument is "HTML", then results are in html format.
}
