% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sas_session_ssh}
\alias{sas_session_ssh}
\title{create sas_session based on configuration file}
\usage{
sas_session_ssh(sascfg = "sascfg_personal.py")
}
\arguments{
\item{sascfg}{(\code{character})\cr SAS session condifuration}
}
\description{
create sas_session based on configuration file
}
