% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sas_session_ssh}
\alias{sas_session_ssh}
\title{Create sas_session based on configuration file}
\usage{
sas_session_ssh(sascfg = get_sas_cfg())
}
\arguments{
\item{sascfg}{(\code{character})\cr SAS session configuration.}
}
\description{
Create sas_session based on configuration file
}
