% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sasr-package.R
\docType{package}
\name{sasr-package}
\alias{sasr-package}
\alias{_PACKAGE}
\title{\code{sasr} Package}
\description{
\code{sasr} provides interface to SAS through \code{saspy} and \code{reticulate} in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/insightsengineering/sasr/}
  \item Report bugs at \url{https://github.com/insightsengineering/sasr/issues}
}

}
\author{
\strong{Maintainer}: insightsengineering \email{insightsengineering@example.com}

}
