% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sasr.R
\name{sasr}
\alias{sasr}
\title{Convert SAS code into R code}
\usage{
sasr(text)
}
\arguments{
\item{text}{the SAS code whatever the indentation}
}
\value{
The R code generated
}
\description{
Simply convert the SAS code into R code. The R code generated
use the following R packages: dplyr, magrittr, readr
}
\examples{

sasr("
proc append base=dataset data=data_to_append getsort force;
run;
")

sasr("
proc compare base=dataset1 compare=adtaset2 printall;
run;

proc compare base=dataset1 compare=adtaset2 printall;
var var1;
with var2;
run;
")

sasr("
proc sort data=dataset out=sorted_dataset;
by var1 descending var2;
run;
")

sasr("
data dataset;
rename var1=new_var1
var2-var3=new_var2-new_var3;
run;
")

sasr("
data new;
set old;
run;
")

sasr("
proc print data=sashelp.iris;
run;
")

sasr("
proc MEANS data=sashelp.iris;
run;
")

sasr("
proc import datafile='c:/temp/test.txt'
out=dataset
dbms=dlm
replace;
delimiter='09'x;
datarow=5;
run;
")

sasr("
proc import datafile='c:/temp/test.csv'
out=dataset
dbms=csv
replace;
getnames=Yes;
run;
")

sasr("
proc transpose data=dataset out=dataset_transposed;
run;
")

sasr("
proc reg data=dataset;
model y = x1 x2;
run;
")

}
