% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sas.R
\name{sd2df}
\alias{sd2df}
\title{Transfer SAS data to R}
\usage{
sd2df(table, libref = "", ..., sas_session = get_sas_session())
}
\arguments{
\item{table}{(\code{character})\cr table name in SAS.}

\item{libref}{(\code{character})\cr library name in SAS.}

\item{...}{additional arguments for \code{saspy.sasbase.SASsession.sd2df}}

\item{sas_session}{(\code{saspy.sasbase.SASsession}) SAS session.}
}
\description{
Transfer SAS data to R
}
