% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_sascfg}
\alias{validate_sascfg}
\title{validate sas configuration file is valid}
\usage{
validate_sascfg(sascfg)
}
\arguments{
\item{sascfg}{(\code{character})\cr file path of configuration.}
}
\description{
validate sas configuration file is valid
}
\details{
Currently, only the file existence check is conducted and the rest
is checked at python side.
}
