% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_ssh_with_tunnel}
\alias{validate_ssh_with_tunnel}
\title{Validate the SAS ssh Session Has Tunnels}
\usage{
validate_ssh_with_tunnel(
  session,
  msg = "SAS session through ssh must use tunnels to transfer datasets!"
)
}
\arguments{
\item{session}{(\code{saspy.sasbase.SASsession}) SAS session.}

\item{msg}{(\code{character})\cr message to display.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Validate if the SAS session has tunnels enabled.

SAS session must enable tunnels to transfer datasets. If not used, a error will pop up.
}
\keyword{internal}
